requirements.txt
omnivoice
static-ffmpeg
pywebview
